[Added] 로또번호 정렬 테스트 작성 및 구현하다.
